// src/generated/prisma/enums.ts
export const PaymentMethod = {
  CASH: 'CASH',
  TRANSFER_BANK: 'TRANSFER_BANK',
  QRIS: 'QRIS',
  COD: 'COD',
  KREDIT: 'KREDIT',
  DEBIT_CARD: 'DEBIT_CARD',
  CREDIT_CARD: 'CREDIT_CARD'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]